backend/requirements.txt
pymupdf>=1.24.0
pillow>=10.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
watchdog>=4.0.0
python-multipart>=0.0.9
